+++ src/components/emtLevel/emtLevel.js
@@ -1,14 +1,7 @@
 import React from "react"
 import PropTypes from "prop-types"
-import { css } from "@emotion/core"
-import Screen from "../utils/screen.js"
 
-const colors = {
-  "first-responder": "deepskyblue",
-  basic: "green",
-  advanced: "#ffbc0d",
-  paramedic: "red",
-}
+import * as styles from "./styles.js"
 
 const capitalize = level => {
   const firstLetter = level[0].toUpperCase()
@@ -16,42 +9,6 @@ const capitalize = level => {
 
   return firstLetter + restOfWord
 }
-
-const emtHeaderCSS = level => css`
-  margin: 0;
-  font-size: 1.2rem;
-  color: ${colors[level]};
-
-  ${Screen.small} {
-    font-size: 1.3rem;
-  }
-`
-
-const linkAnchorCSS = level => css`
-  color: ${colors[level]};
-  text-decoration: none;
-
-  &:hover {
-    text-decoration: underline;
-  }
-`
-
-const emtLevelCSS = level => css`
-  margin: 0.5rem auto 2rem auto;
-  padding: 1.5rem;
-  border: 2px solid ${colors[level]};
-  border-radius: 12px;
-  font-size: 1.15rem;
-
-  ul {
-    padding-left: 1rem;
-    padding-right: 0.5rem;
-  }
-
-  ${Screen.small} {
-    font-size: 1.25rem;
-  }
-`
 
 const EmtLevel = ({ children, level, medControl }) => {
   const standingOrders = " (Standing Orders)"
@@ -110,12 +67,12 @@ const EmtLevel = ({ children, level, medControl }) => {
   }
   return (
     <>
-      <h2 id={getLevel()} css={emtHeaderCSS(level)}>
-        <a css={linkAnchorCSS(level)} href={`#${getLevel()}`}>
+      <h2 id={getLevel()} css={styles.header(level)}>
+        <a css={styles.link(level)} href={`#${getLevel()}`}>
           {getLevelText()}
         </a>
       </h2>
-      <section className={level} css={emtLevelCSS(level)}>
+      <section className={level} css={styles.level(level)}>
         {children}
       </section>
     </>
