+++ src/components/informationBox/informationBox.js
@@ -1,30 +1,7 @@
 import React from "react"
-import { css } from "@emotion/core"
-import Screen from "../utils/screen.js"
 
-import * as styles from "../styles/global.js"
-
-const infoCSS = ({ alignment = "center", borderColor }) => {
-  return css`
-    margin: 0.75rem;
-    padding: 1.5rem 2rem;
-    border-radius: 12px;
-    border: 2px solid ${borderColor};
-    text-align: ${alignment};
-    font-size: 1rem;
-
-    ${Screen.small} {
-      font-size: 1.1rem;
-    }
-  `
-}
-
-const textCSS = ({ alignment, borderColor }) => css`
-  margin: 1rem 0 -0.25rem 0;
-  padding-left: 1rem;
-  text-align: ${alignment};
-  color: ${borderColor};
-`
+import * as styles from "../../styles/global.js"
+import * as infoStyles from "./styles.js"
 
 const InformationBox = ({
   id,
@@ -38,14 +15,18 @@ const InformationBox = ({
   const IdAnchor = () => (
     <a
       href={`#${id}`}
-      css={[styles.links, textCSS({ alignment, borderColor })]}
+      css={[styles.links, infoStyles.text({ alignment, borderColor })]}
     >
       <strong>{text}</strong>
     </a>
   )
 
   const InfoBox = () => (
-    <div id={id} css={[infoCSS({ alignment, borderColor }), css]} {...rest}>
+    <div
+      id={id}
+      css={[infoStyles.info({ alignment, borderColor }), css]}
+      {...rest}
+    >
       {children}
     </div>
   )
