+++ src/components/nextPage/styles.js
@@ -1,0 +1,7 @@
+import { css } from "@emotion/core"
+
+export const nextPage = css`
+  text-align: right;
+  padding: 0;
+  margin: 0;
+`
