+++ src/components/prevPage/styles.js
@@ -1,0 +1,7 @@
+import { css } from "@emotion/core"
+
+export const prevPage = css`
+  text-align: left;
+  padding: 0;
+  margin: 0;
+`
