+++ src/components/title/styles.js
@@ -1,0 +1,6 @@
+import { css } from "@emotion/core"
+
+export const title = css`
+  margin-top: 2rem;
+  text-align: center;
+`
