+++ src/components/title/title.js
@@ -1,13 +1,8 @@
 import React from "react"
 import { useCurrentDoc } from "docz"
-import { css } from "@emotion/core"
 
-import * as styles from "../styles/global.js"
-
-const titleCSS = css`
-  margin-top: 2rem;
-  text-align: center;
-`
+import * as styles from "../../styles/global.js"
+import * as titleStyles from "./styles.js"
 
 const Title = ({ title, ...rest }) => {
   if (!title) {
@@ -19,7 +14,7 @@ const Title = ({ title, ...rest }) => {
   }
 
   return (
-    <h2 id={title} css={titleCSS} {...rest}>
+    <h2 id={title} css={titleStyles.title} {...rest}>
       <a href={`#${title}`} css={styles.links}>
         {title}
       </a>
